% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencies.R
\name{blockEfficiencies}
\alias{blockEfficiencies}
\title{Efficiency factors}
\usage{
blockEfficiencies(TF, BF, treatments_model = NULL)
}
\arguments{
\item{TF}{the treatments factor data frame}

\item{BF}{the block factors data frame}

\item{treatments_model}{the treatments model formula for the treatments data frame. Default is a full factorial model.}
}
\description{
Finds efficiency bounds for block designs.
}
\details{
efficiency factors of regular block designs
}
