% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildblocks.R
\name{buildblocks}
\alias{buildblocks}
\title{buildblocks}
\usage{
buildblocks(TF, blkDesign, searches, seed, jumps)
}
\arguments{
\item{TF}{is a treatment factor with a treatment factor level for each plot level}

\item{blkDesign}{is a data frame with a column for each set of nested blocks and a row for each plot}

\item{searches}{is the number of optimizations searched}
}
\description{
Internal function for optimizing nested block designs for a given set of nested block levels 
\code{blkDesign} and a required set of treatments \code{TF}
}
\keyword{internal}
