% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{durban}
\alias{durban}
\title{Durban example data design}
\format{
An object of class \code{data.frame} with 544 rows and 5 columns.
}
\usage{
data(durban)
}
\description{
Actual layout used by DURBAN, M., HACKETT, C., MCNICOL, J., NEWTON, A., THOMAS, W., & CURRIE, I. (2003). 
The practical use of semi-parametric models in field trials, Journal of Agric Biological and Envir Stats, 8, 48-66.
This data set was obtained from the durban.rowcol {agridat} example in package 'agridat' 
Kevin Wright (2020). agridat: Agricultural Datasets. R package version 1.17.
}
\keyword{data}
