% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Efficiencies.R
\name{EstEffics}
\alias{EstEffics}
\title{D-Efficiency factors}
\usage{
EstEffics(TF, BF)
}
\arguments{
\item{TF}{a treatments factor data frame}

\item{BF}{a block factors data frame}
}
\description{
Finds D and A-efficiency for an unstructured treatmnent set TF with blocks factor BF
}
\details{
efficiency factors of regular block designs
}
