% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcf.R
\name{HCF}
\alias{HCF}
\title{Find hcf}
\usage{
HCF(v)
}
\arguments{
\item{v}{is the vector of integers for which the hcf is required (must be integers)}
}
\value{
hcf
}
\description{
Finds the highest common factor (hcf) of a set of integer numbers greater than zero (Euclidean algorithm)
}
\details{
Finds the hcf of a vector of positive integers which can be in any order
}
\examples{

# hcf of vectors of integers
HCF(c(56,77,616))
HCF(c(3,56,77,616))  

}
