% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPrime.R
\name{isPrime}
\alias{isPrime}
\title{Prime number test}
\usage{
isPrime(v)
}
\arguments{
\item{v}{is the number to be tested for primality}
}
\value{
logical TRUE or FALSE
}
\description{
Tests if a given number v is prime and returns TRUE or FALSE
}
\details{
primality of any positive integer based on the fact that all primes except 2 and 3 can be expressed as 6k-1 or 6k+1
}
\examples{

isPrime(731563)
isPrime(7315631)
 
}
