% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPrmePower.R
\name{isPrimePower}
\alias{isPrimePower}
\title{Prime number test}
\usage{
isPrimePower(N)
}
\arguments{
\item{N}{is the number to be tested for primality}
}
\value{
logical TRUE or FALSE
}
\description{
Tests if a given number v is prime power and returns the base prime and the power or -1
}
\details{
tests if a given number N gives a prime Q when raised to the power Q=N**1/i for all i such that Q>1
}
\examples{

isPrimePower(10000)
 
}
