% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockEfficiencies.R
\name{blockEfficiencies}
\alias{blockEfficiencies}
\title{D-Efficiency factors}
\usage{
blockEfficiencies(TF, BF, treatments_model = NULL)
}
\arguments{
\item{TF}{the treatments factor data frame}

\item{BF}{the block factors data frame}

\item{treatments_model}{a model formula for the required treatments design where the default 
formula assumes a fully crossed factorial treatment model.}
}
\description{
Finds D-efficiency for general treatment and block designs.
}
\details{
efficiency factors of regular block designs
}
