% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcf.R
\name{HCF}
\alias{HCF}
\title{Finds hcf of any set of positive integers}
\usage{
HCF(...)
}
\arguments{
\item{...}{any set of positive integers, in any order, for which the hcf is required.}
}
\value{
hcf
}
\description{
Finds the highest common factor (hcf) of a set of integer numbers greater than zero (Euclidean algorithm).
}
\details{
Finds the hcf of any set of positive integers which can be in any order.
}
\examples{

# hcf of vectors of integers
HCF(56,77,616)
HCF(3,56,77,616) 

}
