% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPrime.R
\name{isPrime}
\alias{isPrime}
\title{Prime number test}
\usage{
isPrime(v)
}
\arguments{
\item{v}{the number to be tested for primality}
}
\value{
logical TRUE or FALSE
}
\description{
Tests if a given number is prime and returns TRUE or FALSE
}
\details{
Tests for the primality of any positive integer using the fact that all primes except 2 and 3 can be
expressed as 6k-1 or 6k+1 for integer k.
}
\examples{

isPrime(731563)
isPrime(7315631)
isPrime(31**2)
 
}
