% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPrmePower.R
\name{isPrimePower}
\alias{isPrimePower}
\title{Finds a prime power solution for N, if available.}
\usage{
isPrimePower(N)
}
\arguments{
\item{N}{the number to be tested for primality}
}
\value{
Returns the base prime p and the power q if N is a prime power; otherwise returns p = 0  and q = 0.
}
\description{
Tests if a given number N is a prime power and returns either the base prime p and power q 
or p = 0 and q = 0.
}
\details{
Finds the smallest integral solution for s = N**(1/i), which gives the smallest s such that
s**i = N. Then, if s is a prime, the number N is a prime power with p = s and q = i.
}
\examples{

isPrimePower(10000)
 
}
